fix: make socket more race-resilient, make tests more readable
While debugging with @fylock we noticed a certain race happening when
the socket first sets up if the remote is not quite ready for data yet. The
`ready` calls were supposed to solve this, but they didn't do so in all case.
This fixes that; "ready" is now a full request/response. Both sides
send it when they're ready. The first person to send their ready will
generally not receive a reply because the other page is still loading.
The second person will. Now, the "ready" reset happens both when the frame
receives a ready call *and* its reply.

I also refactored tests significantly. They were becoming extremely hard to
follow and write previously; I took inspiration from rxjs marbles and created
a sequence-based test runner. The test runner code is ugly (whatever) but
the tests are way easier to read, write, and reason about now.